+++ packages/neuron-ui/src/components/LockInfoDialog/index.tsx
@@ -1,9 +1,8 @@
 import { useTranslation } from 'react-i18next'
 import React, { useRef } from 'react'
-import { CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
-import { scriptToAddress } from 'utils/scriptAndAddress'
+import { type CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
 import Dialog from 'widgets/Dialog'
-import { useCopy, useDialog } from 'utils'
+import { useCopy, useDialog, scriptToAddress } from 'utils'
 import { Copy } from 'widgets/Icons/icon'
 import Alert from 'widgets/Alert'
 import getLockSupportShortAddress from '../../utils/getLockSupportShortAddress'
