+++ packages/neuron-ui/src/components/NervosDAODetail/CellsCard/index.tsx
@@ -1,9 +1,8 @@
 import React, { FC, useMemo, useState } from 'react'
 import Tabs, { VariantProps } from 'widgets/Tabs'
-import { CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
-import { clsx, localNumberFormatter, shannonToCKBFormatter, isMainnet as isMainnetUtils } from 'utils'
+import { type CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
+import { clsx, localNumberFormatter, shannonToCKBFormatter, scriptToAddress, isMainnet as isMainnetUtils } from 'utils'
 import { useTranslation } from 'react-i18next'
-import { scriptToAddress } from 'utils/scriptAndAddress'
 import { onEnter } from 'utils/inputDevice'
 import { EyesClose, EyesOpen } from 'widgets/Icons/icon'
 import { HIDE_BALANCE } from 'utils/const'
