+++ packages/neuron-ui/src/components/Send/hooks.ts
@@ -1,6 +1,6 @@
 import React, { useState, useCallback, useEffect, useMemo } from 'react'
 import { TFunction } from 'i18next'
-import { CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
+import { type CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
 import { AppActions, StateDispatch } from 'states/stateProvider/reducer'
 import { generateTx, generateSendingAllTx } from 'services/remote/wallets'
 import { ControllerResponse, SuccessFromController } from 'services/remote/remoteApiWrapper'
@@ -9,12 +9,12 @@ import {
   outputsToTotalAmount,
   CKBToShannonFormatter,
   shannonToCKBFormatter,
+  scriptToAddress,
   calculateFee,
   validateOutputs,
   DefaultLockInfo,
   RoutePath,
 } from 'utils'
-import { scriptToAddress } from 'utils/scriptAndAddress'
 import { PlaceHolderArgs } from 'utils/const'
 import { useNavigate } from 'react-router-dom'
