+++ packages/neuron-ui/src/utils/getLockSupportShortAddress.ts
@@ -1,4 +1,4 @@
-import { CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
+import { type CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
 import { AnyoneCanPayLockInfoOnAggron, AnyoneCanPayLockInfoOnLina, DefaultLockInfo, MultiSigLockInfo } from './enums'
 
 const getLockSupportShortAddress = (lock: CKBComponents.Script) => {
