+++ packages/neuron-ui/src/utils/getMultisigSignStatus.ts
@@ -1,4 +1,4 @@
-import { addressToScript } from 'utils/scriptAndAddress'
+import { addressToScript } from 'utils'
 import { computeScriptHash } from '@ckb-lumos/base/lib/utils'
 import { MultisigConfig } from 'services/remote'
