+++ packages/neuron-ui/src/utils/outPointTransform.ts
@@ -1,4 +1,4 @@
-import { CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
+import { type CKBComponents } from '@ckb-lumos/rpc/lib/types/api'
 
 export const outPointToStr = (value: CKBComponents.OutPoint): string => {
   return `${value.txHash}_${value.index}`
